refactor(ble): split handler-mqtt-proxy.ts into focused module directory (#131)
Replace the 945-line src/ble/handler-mqtt-proxy.ts with a directory of
seven focused modules. Public API and behaviour are unchanged: same
exported names, same MQTT topic layout, same reconnect/dedup/grace
semantics.

  handler-mqtt-proxy/
  ├── index.ts    public re-exports
  ├── topics.ts   topics() helper, COMMAND_TIMEOUT_MS, Topics type
  ├── client.ts   MQTT client lifecycle, private proxyState, accessors
  ├── gatt.ts     MqttBleChar/Device, mqttGattConnect/Disconnect
  ├── scan.ts     scanAndReadRaw/Read/Devices, mqttScan, waitForEsp32Online
  ├── watcher.ts  ReadingWatcher (event-driven continuous mode)
  └── display.ts  publishConfig/Beep/DisplayReading/Result, registerScaleMac

proxyState (persistent client, discovered MACs, display users) now lives
exclusively in client.ts. Other modules go through accessor helpers
(getClient, hasDiscoveredMac, getDisplayUsers, ...) instead of importing
the raw object, so the shared mutable state is explicit and resettable.

Five import sites updated to point at handler-mqtt-proxy/index.js
(Node ESM does not implicit-resolve directory imports). CONTRIBUTING.md
file system map refreshed.
